sudo-er check

+++ make-blocklist.sh
@@ -1,5 +1,14 @@
 #!/bin/bash
 
+
+### check for root privilges ###
+if [ "$EUID" -ne 0 ]
+then
+  echo -e "\e[91mPlease run as root.\e[39m Root privileges are needed to move and edit files in /etc/pihole"
+  exit
+fi
+
+
 ### reset files ###
 sudo rm -rf ./.blocklist-work*
 sudo rm -rf ./blocklist-fin
